feat: split js file on components
